requirements.txt
fastapi
uvicorn
pydantic
requests
ccxt
web3
python-dotenv
python-telegram-bot
httpx
supabase
